requirements.txt
absl-py==0.8.1
altair==4.0.0
appdirs==1.4.3
argh==0.26.2
aspy.yaml==1.3.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==1.0.3
black==19.10b0
bleach==3.1.0
blinker==1.4
boto3==1.10.37
botocore==1.13.37
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
cfgv==2.0.1
chainer==7.0.0
chardet==3.0.4
Click==7.0
configparser==4.0.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Deprecated==1.2.7
docker-pycreds==0.4.0
docutils==0.15.2
dominate==2.4.0
-e git+https://github.com/hperrot/edflow.git@6a4fa0d3c75b21dfdadb15819dae86794caa3ba8#egg=edflow
Eel==0.12.0
entrypoints==0.3
enum-compat==0.0.3
fastnumbers==2.2.1
filelock==3.0.12
-e git+https://github.com/hperrot/flowiz@b20fa7ad41f02ee69b1db8cd21acc728e1490d97#egg=flowiz
-e git+https://github.com/hperrot/Forward-Warp.git@d9cd57b6e00e4ea88719b59d9757809be021786e#egg=Forward_Warp
future==0.18.2
gevent==1.4.0
gevent-websocket==0.10.1
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.10.0
google-auth-oauthlib==0.4.1
gql==0.1.0
graphql-core==2.2.1
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
identify==1.4.9
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
importlib-resources==1.0.2
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
kornia==0.1.4.post2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
natsort==6.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
ninja==1.9.0.post1
nodeenv==1.3.3
nose==1.3.7
notebook==6.0.2
numpy==1.17.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.6.0
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==6.2.1
plotly==4.4.1
pluggy==0.13.1
pre-commit==1.20.0
prettytable==0.7.2
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.11.1
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
regex==2019.12.9
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.3
Send2Trash==1.5.0
sentry-sdk==0.13.5
shortuuid==0.5.0
six==1.13.0
sklearn==0.0
smmap2==2.0.5
streamlit==0.51.0
subprocess32==3.5.4
tensorboard==2.1.0
tensorboardX==1.9
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
torch
torchfields==0.0.5
torchvision
tornado==5.1.1
tqdm==4.40.2
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.1
virtualenv==16.7.9
wandb==0.8.18
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
